Adds segmentation model to label frames and ability to learn from them via wrappers (#6)
* Creates rudimentary dataset

* Adds function for randomly cropping and flipping images

* Casts to tensor

* Adds class for data segments

* Adds class for data segments

* Adds support for train & eval

* Adds utils for segmentation

* Adds utils for segmentation

* Removes deprecated function from imports

* Separates paths for images and labels

* Adds (v0) of segmentation model inside make_env

* Corrects type in model loading

* Adds temp path for segmentation model

* Corrects syntax for torch.load

* Corrects typo in ProcessFrame84Segment

* Casts frame to device

* Makes copy of image for numpy negative strides

* Scales segments between 0 and 255 (most straightforward for testing)

* Adds reference for compatibility with lightning

* Corrects type annotation for float

+++ agent/ac.py
@@ -32,7 +32,7 @@ class ActorCritic(nn.Module):
     def __init__(
         self,
         # ~ Env. variables ~
-        state_space: npt.NDArray[np.float64],
+        state_space: npt.NDArray[np.float32],
         action_space: int,
         # ~ Actor variables ~
         policy_lr: float,
